Added parcel support and more
also fixed sprite bug
