+++ src/www/simple/components/App.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+class App extends React.Component {
+    render() {
+        const { store } = this.props;
+        return (
+            <div>Hello, from {store.config.code}</div>
+        );
+    }
+}
